Reimplemented member attending counter.
